+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 1.0.15
-*
+* Add timeout option
 
 ### 1.0.14
 * Switch from deleteAgent to postAgent operation for all agents delete action